put defaults in metadata so they are overwritable
when defaults are set as variables in defaults.yaml, they are considered command line "defaults" and thus are not overwritable by values in the document yaml header, which is where we would want to change behavior specific to that document. By including these values in metadata.yaml (which is then read in by defaults.yaml) the values can be overwritten by document yaml header.